FIX: NTP Issues, SMB issues (#1073)
Disabled guest SMB access as it was causing issues because [Microsoft
disabled guest SMB login per
default](https://learn.microsoft.com/en-us/troubleshoot/windows-server/networking/guest-access-in-smb2-is-disabled-by-default)

change curl timeouts from --connect-timeout to --max-time (-m) 

Fix the endless wait on startup when no IP address was assigned.
_Someone_ <sub><sup><sub><sup><sub>(me)</sub></sup></sub></sup></sub>
forgot to break out of the loop.
Add a check to make sure update_networking isn't run multiple times
simultaneously.

TESTS:

- [ ] Enable SMB, connect to it. Observe that it asks for credentials
(unless you've saved them)
- [ ] Enable NTP and wait for NTP on boot
- [ ] reboot the device on functioning network. observe the time syncs
- [ ] make your internet connection not working (e.g. pull dsl cable)
but the wifi still available
- [ ] reboot the device, observe that the sync times out in a reasonable
time
- [ ] disable your wifi all together (or pull the router plug etc)
- [ ] reboot the device again, observe that the sync times out in a
reasonable time

---------

+++ static/build/.tmp_update/runtime.sh
@@ -125,6 +125,7 @@ main() {
 }
 
 state_change() {
+    log "state change: $1"
     runifnecessary "keymon" keymon
     check_networking
     touch /tmp/state_changed
@@ -583,9 +584,12 @@ check_networking() {
         return
     fi
 
-    rm /tmp/network_changed
-
-    $sysdir/script/network/update_networking.sh check
+    if pgrep -f update_networking.sh; then
+        log "update_networking already running"
+    else
+        rm /tmp/network_changed
+        $sysdir/script/network/update_networking.sh check
+    fi
 
     check_timezone
 }
